apply DRY and refactor code
