--- decidim-accountability/app/cells/decidim/accountability/result_history/show.erb
@@ -1,25 +1,0 @@
-<ul class="result_history_cell">
-  <% history_items.each_with_index do |item, index| %>
-    <li class="result_history_cell__item">
-      <div class="result_history_cell__item_icon">
-        <div class="<%= item[:id] == "result_creation" ? "result_history_cell__item_icon-creation" : "" %>">
-          <%= icon item[:icon], class: item[:id] == "result_creation" ? "fill-current text-white" : "fill-current text-secondary" %>
-        </div>
-        <% if index < history_items.size - 1 %>
-          <span class="<%= item[:id] == "result_creation" ? "result_history_cell__line_short" : "result_history_cell__line_long" %>"></span>
-        <% end %>
-      </div>
-      <div class="result_history_cell__item_content">
-        <span class="result_history_cell__item_date"><%= item[:date].strftime("%d/%m/%Y %H:%M") %></span>
-        <p><%= item[:text] %></p>
-        <% if item[:resource].present? %>
-          <ul>
-            <li>
-              <%= link_to decidim_sanitize_translated(item[:resource].title), item[:url], class: "underline decoration-current text-secondary font-semibold" %>
-            </li>
-          </ul>
-        <% end %>
-      </div>
-    </li>
-  <% end %>
-</ul>
