+++ decidim-accountability/app/packs/entrypoints/decidim_accountability.js
@@ -5,4 +5,3 @@ require.context("../images", true)
 
 // CSS
 import "stylesheets/accountability.scss"
-import "stylesheets/result_history.scss"
