--- decidim-budgets/app/cells/decidim/budgets/budget_history/show.erb
@@ -1,20 +1,0 @@
-<ul class="budget_history_cell">
-  <% history_items.each_with_index do |item, index| %>
-    <li class="budget_history_cell__item">
-      <div class="budget_history_cell__item_icon">
-        <div class="<%= item[:id] == "budget_creation" ? "budget_history_cell__item_icon-creation" : "" %>">
-          <%= icon item[:icon], class: item[:id] == "budget_creation" ? "fill-current text-white" : "fill-current text-secondary" %>
-        </div>
-        <% if index < history_items.size - 1 %>
-          <span class="<%= item[:id] == "budget_creation" ? "budget_history_cell__line_short" : "budget_history_cell__line_long" %>"></span>
-        <% end %>
-      </div>
-      <div class="budget_history_cell__item_content">
-        <span class="budget_history_cell__item_date"><%= item[:date].strftime("%d/%m/%Y %H:%M") %></span>
-        <p class="flex flex-col lg:flex-none lg:block">
-         <%= item[:text] %>
-        </p>
-      </div>
-    </li>
-  <% end %>
-</ul>
