+++ decidim-budgets/app/cells/decidim/budgets/budget_history_cell.rb
@@ -14,42 +14,15 @@ module Decidim
     #     "decidim/budget_history",
     #     budget
     #   )
-    class BudgetHistoryCell < Decidim::ViewModel
+    class BudgetHistoryCell < Decidim::ResourceHistoryCell
       include Decidim::Budgets::ApplicationHelper
-
-      def show
-        render
-      end
 
       private
 
-      def history_items
-        return @history_items if @history_items.present?
-
-        @history_items = []
+      def add_history_items
         add_linked_resources_items(@history_items, :proposals, "included_proposals", "decidim/proposals/proposal/budget_text", "Decidim::Proposals::Proposal")
         add_linked_resources_items(@history_items, :results, "included_projects", "decidim/accountability/result/budget_text", "Decidim::Accountability::Result")
         add_budget_creation_item(@history_items) if @history_items.any?
-
-        @history_items.sort_by! { |item| item[:date] }
-      end
-
-      def add_linked_resources_items(items, resource_type, link_name, text_key, icon_key)
-        resources = @model.linked_resources(resource_type, link_name)
-        return if resources.blank?
-
-        resources.each do |resource|
-          title = decidim_sanitize_translated(resource.title[I18n.locale.to_s])
-          url = resource_locator(resource).path
-          link = link_to(title, url, class: "underline decoration-current text-secondary font-semibold")
-
-          items << {
-            id: "#{link_name}_#{resource.id}",
-            date: resource.updated_at,
-            text: t(text_key, scope: "activerecord.models", count: 1, link:),
-            icon: resource_type_icon_key(icon_key)
-          }
-        end
       end
 
       def add_budget_creation_item(items)
@@ -59,6 +32,10 @@ module Decidim
           text: t("decidim.budgets.creation.text"),
           icon: resource_type_icon_key("Decidim::Budgets::Project")
         }
+      end
+
+      def history_cell_id
+        "budget"
       end
     end
   end
