+++ decidim-budgets/app/packs/entrypoints/decidim_budgets.js
@@ -6,4 +6,3 @@ import "src/decidim/budgets/exit_handler"
 require.context("../images", true)
 
 import "stylesheets/budgets.scss"
-import "stylesheets/budget_history.scss"
