+++ decidim-core/app/packs/entrypoints/decidim_core.js
@@ -6,6 +6,7 @@ import "src/decidim/decidim_application"
 
 // CSS
 import "entrypoints/decidim_core.scss"
+import "stylesheets/decidim/resource_history.scss"
 
 // Images
 require.context("../images", true)
