+++ decidim-core/app/packs/stylesheets/decidim/resource_history.scss
@@ -1,4 +1,4 @@
-.result_history_cell {
+.resource_history {
   @apply relative flex flex-col gap-4 p-4;
 
   &__line_short {
@@ -7,6 +7,10 @@
 
   &__line_long {
     @apply absolute left-1/2 top-full w-px h-11 lg:h-4 bg-secondary;
+  }
+
+  &__line_large {
+    @apply absolute left-1/2 top-full w-px h-16 lg:h-4 bg-secondary;
   }
 
   &__item {
