+++ decidim-proposals/app/cells/decidim/proposals/proposal_history_cell.rb
@@ -14,19 +14,12 @@ module Decidim
     #     "decidim/proposal_history",
     #     proposal
     #   )
-    class ProposalHistoryCell < Decidim::ViewModel
+    class ProposalHistoryCell < Decidim::ResourceHistoryCell
       include Decidim::Proposals::ApplicationHelper
-
-      def show
-        render
-      end
 
       private
 
-      def history_items
-        return @history_items if @history_items.present?
-
-        @history_items = []
+      def add_history_items
         # linked resources generate from this proposal
         resources = @model.linked_resources_from(:proposals, "copied_from_component")
         add_linked_resources_items(@history_items, resources, "copied_from_component", "decidim/proposals/proposal/import_from_proposal_text", "Decidim::Proposals::Proposal")
@@ -51,27 +44,6 @@ module Decidim
         add_linked_resources_items(@history_items, resources, "merged_to_component", "decidim/proposals/proposal/merge_to_proposal_text", "Decidim::Proposals::Proposal")
 
         add_proposal_creation_item(@history_items) if @history_items.any?
-
-        @history_items.sort_by! { |item| item[:date] }
-      end
-
-      def add_linked_resources_items(items, resources, link_name, text_key, icon_key)
-        return if resources.blank?
-
-        resources.each do |resource|
-          if text_key == "decidim/proposals/proposal/import_to_proposal_text"
-            link = link_to(decidim_sanitize_translated(resource.title[I18n.locale.to_s]), resource_locator(resource).path,
-                           class: "underline decoration-current text-secondary font-semibold")
-          end
-          items << {
-            id: "#{link_name}_#{resource.id}",
-            date: resource.updated_at,
-            text: link ? t(text_key, scope: "activerecord.models", link:) : t(text_key, scope: "activerecord.models"),
-            icon: resource_type_icon_key(icon_key),
-            url: resource_locator(resource).path,
-            resource:
-          }
-        end
       end
 
       def add_proposal_creation_item(items)
@@ -92,10 +64,8 @@ module Decidim
         }
       end
 
-      def history_items_contains?(link_name)
-        return false if @history_items.blank?
-
-        @history_items.any? { |item| item[:id].include?(link_name.to_s) }
+      def history_cell_id
+        "proposal"
       end
     end
   end
