+++ decidim-proposals/app/packs/entrypoints/decidim_proposals.js
@@ -10,4 +10,3 @@ require.context("../documents", true)
 
 // CSS
 import "stylesheets/decidim/proposals/proposals.scss"
-import "stylesheets/decidim/proposals/proposal_history.scss"
