+++ decidim-proposals/config/locales/en.yml
@@ -91,12 +91,12 @@ en:
       decidim/proposals/voting_enabled_event: Proposal voting enabled
   activerecord:
     models:
-      decidim/accountability/result/text: 'It was added to this result:'
-      decidim/budgets/project/text: 'It was added to this budget:'
+      decidim/accountability/result/text: 'It was added to this result: %{link}'
+      decidim/budgets/project/text: 'It was added to this budget: %{link}'
       decidim/history/history:
         one: History
         other: History
-      decidim/meetings/meeting/text: 'It was discussed in this meeting:'
+      decidim/meetings/meeting/text: 'It was discussed in this meeting: %{link}'
       decidim/proposals/collaborative_draft:
         one: Collaborative draft
         other: Collaborative drafts
@@ -104,12 +104,12 @@ en:
         one: Proposal
         other: Proposals
       decidim/proposals/proposal/budget_text: The proposal %{link} was created
-      decidim/proposals/proposal/import_from_proposal_text: 'It became this proposal:'
+      decidim/proposals/proposal/import_from_proposal_text: 'It became this proposal: %{link}'
       decidim/proposals/proposal/import_to_proposal_text: This proposal %{link} was created
-      decidim/proposals/proposal/merge_from_proposal_text: 'It became this proposal:'
-      decidim/proposals/proposal/merge_to_proposal_text: 'This proposal was created:'
-      decidim/proposals/proposal/split_from_proposal_text: 'It became this proposal:'
-      decidim/proposals/proposal/split_to_proposal_text: 'This proposal was created:'
+      decidim/proposals/proposal/merge_from_proposal_text: 'It became this proposal: %{link}'
+      decidim/proposals/proposal/merge_to_proposal_text: 'This proposal was created: %{link}'
+      decidim/proposals/proposal/split_from_proposal_text: 'It became this proposal: %{link}'
+      decidim/proposals/proposal/split_to_proposal_text: 'This proposal was created: %{link}'
       decidim/proposals/proposal_note:
         one: Note
         other: Notes
